move FitWitRooFit out of bin and create a class with a public header
